Completed MVC changes for Technician.
The MVC for Technician model function is complete, now we just have to
polish it up and change links around.

Lessee MVC should be almost identical to the Technician (it also
inherits from User.)

The Technician MVC should have examples for any changes we need to make
to any of the other models.